feat(core): migra para BigDecimal, ajusta ambiente e inicia versao 0.1.0-BETA

+++ src/main/java/com/concessionaria/gomez/domain/model/Carro.java
@@ -6,6 +6,7 @@ import org.hibernate.annotations.CreationTimestamp;
 
 import javax.persistence.*;
 import javax.validation.constraints.NotBlank;
+import java.math.BigDecimal;
 import java.time.OffsetDateTime;
 
 
@@ -30,7 +31,7 @@ public class Carro {
     private Integer ano;
 
     @Column(nullable = false)
-    private Double compra;
+    private BigDecimal compra;
 
     private Boolean ativo = Boolean.TRUE;
